// client/src/interface/userInterface.ts
// export interface IUser {
//   id: 'juhlee';
//   email: 'juhlee@student.42seoul.kr';
//   firstName: 'Juhae';
//   lastName: 'Lee';
//   picture: 'https://cdn.intra.42.fr/users/juhlee.jpg';
// }

export interface IUser {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  picture: string; // url
}